+++ lib/orderbook/OrderBook.ts
@@ -15,7 +15,7 @@ import { CurrencyInstance, PairInstance, CurrencyFactory } from '../db/types';
 import { Pair, OrderIdentifier, OwnOrder, OrderPortion, OwnLimitOrder, PeerOrder, Order, PlaceOrderEvent,
   PlaceOrderEventType, PlaceOrderResult, OutgoingOrder, OwnMarketOrder, isOwnOrder, IncomingOrder } from './types';
 import { SwapRequestPacket, SwapFailedPacket } from '../p2p/packets';
-import { SwapResult, SwapDeal } from 'lib/swaps/types';
+import { SwapSuccess, SwapDeal } from 'lib/swaps/types';
 import Bluebird from 'bluebird';
 
 interface OrderBook {
@@ -95,14 +95,14 @@ class OrderBook extends EventEmitter {
 
   private bindSwaps = () => {
     if (this.swaps) {
-      this.swaps.on('swap.paid', async (swapResult) => {
-        if (swapResult.role === SwapRole.Maker) {
-          const { orderId, pairId, quantity, peerPubKey } = swapResult;
+      this.swaps.on('swap.paid', async (swapSuccess) => {
+        if (swapSuccess.role === SwapRole.Maker) {
+          const { orderId, pairId, quantity, peerPubKey } = swapSuccess;
 
           // we must remove the amount that was put on hold while the swap was pending for the remaining order
           this.removeOrderHold(orderId, pairId, quantity);
 
-          await this.persistTrade(swapResult.quantity, this.getOwnOrder(swapResult.orderId, swapResult.pairId), undefined, swapResult.rHash);
+          await this.persistTrade(swapSuccess.quantity, this.getOwnOrder(swapSuccess.orderId, swapSuccess.pairId), undefined, swapSuccess.rHash);
           this.removeOwnOrder(orderId, pairId, quantity, peerPubKey);
           this.emit('ownOrder.swapped', { pairId, quantity, id: orderId });
         }
@@ -246,7 +246,7 @@ class OrderBook extends EventEmitter {
 
       return {
         internalMatches: [],
-        swapResults: [],
+        swapSuccesses: [],
         remainingOrder: stampedOrder,
       };
     }
@@ -265,6 +265,17 @@ class OrderBook extends EventEmitter {
     return addResult;
   }
 
+  /**
+   * Places an order in the order book. This method first attempts to match the order with existing
+   * orders by price and initiate swaps for any matches with peer orders. It can be called recursively
+   * for any portions of the order that fail swaps.
+   * @param order the order to place
+   * @param discardRemaining whether to discard any unmatched portion of the order, if `false` the
+   * unmatched portion will enter the order book.
+   * @param onUpdate a callback for when there are updates to the matching and order placement
+   * routine including internal matches, successful swaps, failed swaps, and remaining orders
+   * @param maxTime the deadline in epoch milliseconds for this method to end recursive calls
+   */
   private placeOrder = async (
     order: OwnOrder,
     discardRemaining = false,
@@ -278,106 +289,128 @@ class OrderBook extends EventEmitter {
       this.logger.debug(`placeOrder max time exceeded. order (${JSON.stringify(order)}) won't be fully matched`);
 
       // returning the remaining order to be rolled back and handled by the initial call
-      return Promise.resolve({
+      return {
         internalMatches: [],
-        swapResults: [],
+        swapSuccesses: [],
         remainingOrder: order,
-      });
+      };
     }
 
-    // perform match. maker orders will be removed from the repository
+    // perform matching routine. maker orders that are matched will be removed from the order book.
     const tp = this.getTradingPair(order.pairId);
     const matchingResult = tp.match(order);
 
-    // instantiate the final response object
-    const result: PlaceOrderResult = {
-      internalMatches: [],
-      swapResults: [],
-      remainingOrder: matchingResult.remainingOrder,
+    /** Any portion of the placed order that could not be swapped or matched internally. */
+    let { remainingOrder } = matchingResult;
+    /** Local orders that matched with the placed order. */
+    const internalMatches: OwnOrder[] = [];
+    /** Successful swaps performed for the placed order. */
+    const swapSuccesses: SwapSuccess[] = [];
+
+    /**
+     * The routine for retrying a portion of the order that failed a swap attempt.
+     * @param failedSwapQuantity the quantity of the failed portion to retry
+     */
+    const retryFailedSwap = async (failedSwapQuantity: number) => {
+      this.logger.debug(`repeating matching routine for ${order.id} for failed quantity of ${failedSwapQuantity}`);
+      const orderToRetry: OwnOrder = { ...order, quantity: failedSwapQuantity };
+
+      // invoke placeOrder recursively, append matches/swaps and any remaining order
+      const retryResult = await this.placeOrder(orderToRetry, true, onUpdate, maxTime);
+      internalMatches.push(...retryResult.internalMatches);
+      swapSuccesses.push(...retryResult.swapSuccesses);
+      if (retryResult.remainingOrder) {
+        if (remainingOrder) {
+          remainingOrder.quantity += retryResult.remainingOrder.quantity;
+        } else {
+          remainingOrder = retryResult.remainingOrder;
+        }
+      }
     };
 
-    /** The quantity that was attempted to be swapped but failed. */
-    let failedSwapQuantity = 0;
-
-    // iterate over the matches
-    for (const { maker, taker } of matchingResult.matches) {
+    /**
+     * The routine for handling matches found in the order book. This can be run in parallel
+     * so that all matches, including those which require swaps with peers, can be executed
+     * simultaneously.
+     */
+    const handleMatch = async (maker: Order, taker: OwnOrder) => {
       const portion: OrderPortion = { id: maker.id, pairId: maker.pairId, quantity: maker.quantity };
       if (isOwnOrder(maker)) {
-        // internal match
+        // this is an internal match which is effectively executed immediately upon being found
         this.logger.info(`internal match executed on taker ${taker.id} and maker ${maker.id} for ${maker.quantity}`);
         portion.localId = maker.localId;
-        result.internalMatches.push(maker);
+        internalMatches.push(maker);
         this.emit('ownOrder.filled', portion);
         await this.persistTrade(portion.quantity, maker, taker);
         onUpdate && onUpdate({ type: PlaceOrderEventType.InternalMatch, payload: maker });
       } else {
+        // this is a match with a peer order which cannot be considered executed until after a
+        // successful swap, which is an asynchronous process that can fail for numerous reasons
         if (!this.swaps) {
-          // swaps should only be undefined during integration testing of the order book
-          // for now we treat this case like a swap failure
+          // the swaps module should only be undefined during integration testing of the order book
+          // in this case we treat the swap as if it failed, but without retrying the failed portion
           this.emit('peerOrder.invalidation', portion);
-          failedSwapQuantity += portion.quantity;
-          continue;
+          return;
         }
 
         try {
           this.logger.debug(`matched with peer ${maker.peerPubKey}, executing swap on taker ${taker.id} and maker ${maker.id} for ${maker.quantity}`);
-          const swapResult = await this.executeSwap(maker, taker);
+          const swapSuccess = await this.executeSwap(maker, taker);
 
-          if (swapResult.quantity < maker.quantity) {
+          if (swapSuccess.quantity < maker.quantity) {
             // swap was only partially completed
-            portion.quantity = swapResult.quantity;
-            const rejectedQuantity = maker.quantity - swapResult.quantity;
-            failedSwapQuantity += rejectedQuantity; // add unswapped quantity to failed quantity
-            this.logger.info(`match partially executed on taker ${taker.id} and maker ${maker.id} for ${swapResult.quantity} ` +
+            portion.quantity = swapSuccess.quantity;
+            const rejectedQuantity = maker.quantity - swapSuccess.quantity;
+            this.logger.info(`match partially executed on taker ${taker.id} and maker ${maker.id} for ${swapSuccess.quantity} ` +
               `with peer ${maker.peerPubKey}, ${rejectedQuantity} quantity not accepted and will repeat matching routine`);
+            await retryFailedSwap(rejectedQuantity);
           } else {
             this.logger.info(`match executed on taker ${taker.id} and maker ${maker.id} for ${maker.quantity} with peer ${maker.peerPubKey}`);
           }
-          result.swapResults.push(swapResult);
-          onUpdate && onUpdate({ type: PlaceOrderEventType.SwapSuccess, payload: swapResult });
+          swapSuccesses.push(swapSuccess);
+          onUpdate && onUpdate({ type: PlaceOrderEventType.SwapSuccess, payload: swapSuccess });
         } catch (err) {
-          failedSwapQuantity += portion.quantity;
-          onUpdate && onUpdate({ type: PlaceOrderEventType.SwapFailure, payload: maker });
           this.logger.warn(`swap for ${portion.quantity} failed during order matching, will repeat matching routine for failed swap quantity`);
+          onUpdate && onUpdate({ type: PlaceOrderEventType.SwapFailure, payload: maker });
+          await retryFailedSwap(portion.quantity);
         }
       }
+    };
+
+    // iterate over the matches to be executed in parallel
+    const matchPromises: Promise<void>[] = [];
+    for (const { maker, taker } of matchingResult.matches) {
+      matchPromises.push(handleMatch(maker, taker));
     }
 
-    // if we have swap failures, attempt one retry for all available quantity. don't re-add the maker orders
-    if (failedSwapQuantity > 0) {
-      this.logger.debug(`${failedSwapQuantity} quantity of swaps failed for order ${order.id}, repeating matching routine for failed quantity`);
-      // aggregate failures quantities with the remaining order
-      const remainingOrder: OwnOrder = result.remainingOrder || { ...order, quantity: 0 };
-      remainingOrder.quantity += failedSwapQuantity;
+    // wait for all matches to complete execution, any portions that cannot be executed due to
+    // failed swaps will be added to the remaining order which may be added to the order book.
+    await Promise.all(matchPromises);
 
-      // invoke placeOrder recursively, append matches/swaps and set the consecutive remaining order
-      const remainingOrderResult = await this.placeOrder(remainingOrder, true, onUpdate, maxTime);
-      result.internalMatches.push(...remainingOrderResult.internalMatches);
-      result.swapResults.push(...remainingOrderResult.swapResults);
-      result.remainingOrder = remainingOrderResult.remainingOrder;
-    }
-
-    const { remainingOrder } = result;
     if (remainingOrder && !discardRemaining) {
       this.addOwnOrder(remainingOrder);
       onUpdate && onUpdate({ type: PlaceOrderEventType.RemainingOrder, payload: remainingOrder });
     }
 
-    return result;
+    return {
+      internalMatches,
+      swapSuccesses,
+      remainingOrder,
+    };
   }
 
   /**
    * Executes a swap between maker and taker orders. Emits the `peerOrder.filled` event if the swap
    * succeeds and `peerOrder.invalidation` if the swap fails.
    */
-  public executeSwap = async (maker: PeerOrder, taker: OwnOrder): Promise<SwapResult> => {
+  public executeSwap = async (maker: PeerOrder, taker: OwnOrder): Promise<SwapSuccess> => {
     // make sure the order is in the database before we begin the swap
     await this.repository.addOrderIfNotExists(maker);
     try {
-      const swapResult = await this.swaps!.executeSwap(maker, taker);
+      const swapSuccess = await this.swaps!.executeSwap(maker, taker);
       this.emit('peerOrder.filled', maker);
-      await this.persistTrade(swapResult.quantity, maker, taker, swapResult.rHash);
-      return swapResult;
+      await this.persistTrade(swapSuccess.quantity, maker, taker, swapSuccess.rHash);
+      return swapSuccess;
     } catch (err) {
       this.emit('peerOrder.invalidation', maker);
       // TODO: penalize peer for failed swap? penalty severity should depend on reason for failure
@@ -446,51 +479,61 @@ class OrderBook extends EventEmitter {
    * is not supported or if the order to cancel could not be found.
    * @returns any quantity of the order that was on hold and could not be immediately removed.
    */
-  public removeOwnOrderByLocalId = (localId: string) => {
+  public removeOwnOrderByLocalId = (localId: string, quantityToRemove?: number) => {
     const orderIdentifier = this.localIdMap.get(localId);
     if (!orderIdentifier) {
       throw errors.LOCAL_ID_DOES_NOT_EXIST(localId);
     }
 
     const order = this.getOwnOrder(orderIdentifier.id, orderIdentifier.pairId);
-    if (order.hold) {
-      let remainingHold = order.hold;
-      // we can't remove the entire order as some of it is on hold, start by removing any available portion
-      this.logger.debug(`can't remove local order ${localId} yet because it has a hold of ${order.hold}`);
-      const availableQuantity = order.quantity - order.hold;
-      if (availableQuantity) {
-        this.removeOwnOrder(orderIdentifier.id, orderIdentifier.pairId, availableQuantity);
-      }
+
+    let remainingQuantityToRemove = quantityToRemove || order.quantity;
+
+    if (remainingQuantityToRemove > order.quantity) {
+      // quantity to be removed can't be higher than order's quantity.
+      throw errors.QUANTITY_DOES_NOT_MATCH(remainingQuantityToRemove, order.quantity);
+    }
+
+    const removableQuantity = order.quantity - order.hold;
+    if (remainingQuantityToRemove <= removableQuantity) {
+      this.removeOwnOrder(orderIdentifier.id, orderIdentifier.pairId, remainingQuantityToRemove);
+      remainingQuantityToRemove = 0;
+    } else {
+      // we can't remove the entire amount because of a hold on the order
+      this.removeOwnOrder(orderIdentifier.id, orderIdentifier.pairId, removableQuantity);
+      remainingQuantityToRemove -= removableQuantity;
+
+      const failedHandler = (deal: SwapDeal) => {
+        if (deal.orderId === orderIdentifier.id) {
+          // remove the portion that failed now that it's not on hold
+          const quantityToRemove = Math.min(deal.quantity!, remainingQuantityToRemove);
+          this.removeOwnOrder(orderIdentifier.id, orderIdentifier.pairId, quantityToRemove);
+          cleanup(quantityToRemove);
+        }
+      };
+
+      const paidHandler = (result: SwapSuccess) => {
+        if (result.orderId === orderIdentifier.id) {
+          const quantityToRemove = Math.min(result.quantity, remainingQuantityToRemove);
+          cleanup(quantityToRemove);
+        }
+      };
 
       const cleanup = (quantity: number) => {
-        remainingHold -= quantity;
-        this.logger.debug(`removed hold of ${quantity} on local order ${localId}, ${remainingHold} remaining`);
-        if (remainingHold === 0) {
+        remainingQuantityToRemove -= quantity;
+        this.logger.debug(`removed hold of ${quantity} on local order ${localId}, ${remainingQuantityToRemove} remaining`);
+        if (remainingQuantityToRemove === 0) {
           // we can stop listening for swaps once all holds are cleared
           this.swaps!.removeListener('swap.failed', failedHandler);
           this.swaps!.removeListener('swap.paid', paidHandler);
         }
       };
 
-      const failedHandler = (deal: SwapDeal) => {
-        if (deal.orderId === orderIdentifier.id) {
-          // remove the portion that failed now that it's not on hold
-          this.removeOwnOrder(orderIdentifier.id, orderIdentifier.pairId, deal.quantity!);
-          cleanup(deal.quantity!);
-        }
-      };
-
-      const paidHandler = (result: SwapResult) => {
-        if (result.orderId === orderIdentifier.id) {
-          cleanup(result.quantity);
-        }
-      };
-
       this.swaps!.on('swap.failed', failedHandler);
       this.swaps!.on('swap.paid', paidHandler);
     }
-    this.removeOwnOrder(orderIdentifier.id, orderIdentifier.pairId);
-    return order.hold;
+
+    return remainingQuantityToRemove;
   }
 
   private addOrderHold = (orderId: string, pairId: string, holdAmount: number) => {